+++ scripts/src/app.tsx
@@ -1,24 +1,43 @@
-import Config from "./index";
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
 import FormAdminPage from './admin/FormAdminPage';
 import FormPage from "./form/FormPage";
 import FormStandalone from "./form/FormStandalone/FormStandalone";
 import { BrowserRouter as Router, Route, Switch, Redirect } from "react-router-dom";
 import * as queryString from 'query-string';
 import "./app.scss";
-import "./form.tsx";
+import auth from "./Auth";
+import * as DOMPurify from 'dompurify';
 
+import { Provider } from 'react-redux';
+import { Reducer, createStore, applyMiddleware } from 'redux';
+import { reducers } from './store/index';
+import thunkMiddleware from 'redux-thunk'
+import { connectRouter, routerMiddleware } from 'connected-react-router';
+import { ConnectedRouter } from 'connected-react-router';
+import { createBrowserHistory } from 'history'
 
-let formAdminElement = document.getElementById('ccmt-cff-main') as HTMLElement;
-if (true || formAdminElement) {
-  ReactDOM.render(
-    <div className="ccmt-cff-Wrapper-Bootstrap">
-      <Router>
+// Make all external links in form open in a new tab.
+DOMPurify.addHook('afterSanitizeAttributes', function (node) {
+  // set all elements owning target to target=_blank
+  if ('target' in node) {
+    node.setAttribute('target', '_blank');
+  }
+});
+
+const history = createBrowserHistory()
+const store = createStore(
+  connectRouter(history)(reducers),
+  applyMiddleware(thunkMiddleware, routerMiddleware(history))
+);
+
+const App = () => (
+  <Provider store={store}>
+    <ConnectedRouter history={history}>
+      <div className="ccmt-cff-Wrapper-Bootstrap">
         <Switch>
           <Route path="/:centerSlug/forms/:formId" exact render={(props) => {
             return (
-              <FormStandalone {...props} formId={props.match.params.formId} ENDPOINT_URL={Config.ENDPOINT_URL} />
+              <FormStandalone {...props} formId={props.match.params.formId} ENDPOINT_URL={"abcdefg"} />
             );
           }
           } />
@@ -26,18 +45,13 @@ if (true || formAdminElement) {
             <div className="App FormAdminPage">
               <FormAdminPage
                 {...props}
-                apiEndpoint={Config.ENDPOINT_URL}
                 apiKey={null}
-                federated={Config.federated}
               />
             </div>
           } />
           <Redirect to={`/admin/`} />
         </Switch>
-      </Router>
-
-    </div>,
-    formAdminElement
-  );
-
-}
+      </div>
+    </ConnectedRouter>
+  </Provider>);
+export default App;