+++ scripts/src/store/index.ts
@@ -1,0 +1,7 @@
+import { combineReducers, Dispatch, Reducer } from 'redux';
+
+import authReducer from './auth/reducer';
+
+export const reducers: Reducer = combineReducers({
+  auth: authReducer,
+});